+++ src/AggregateRoot.ts
@@ -34,12 +34,12 @@ export function createAggregateRoot<T>(
     };
   };
 
-  const handle = async (
+  const applyCommand = async (
     entity: IAggregateState<T>,
     command: IDomainCommand,
     services: IServiceRegistry
   ): Promise<IDomainEvent[]> => {
-    const handler = commands[command.name];
+    const commandHandlerMapEntry = commands[command.name];
 
     const events: IDomainEvent[] = [];
     let instance: IPublishableAggregateState<T>;
@@ -56,20 +56,32 @@ export function createAggregateRoot<T>(
 
     instance = { ...entity, publish };
 
-    const generator = handler({ ...entity, publish }, command, services);
+    const handlerList = Array.isArray(commandHandlerMapEntry)
+      ? commandHandlerMapEntry
+      : [commandHandlerMapEntry];
 
-    // Handler method can either call `entity.publish` directly & return void,
-    // or it can be a generator, in which case we need to iterate through all
-    // the published events
-    if (generator !== undefined) {
-      let event = await Promise.resolve(generator.next(instance));
+    const _apply = async (fn: any) => {
+      const generator = await Promise.resolve(fn(instance, command, services));
 
-      while (!event.done) {
-        event = await Promise.resolve(generator.next(instance));
+      // Handler method can either call `entity.publish` directly & return void,
+      // or it can be a generator, in which case we need to iterate through all
+      // the published events
+      if (generator !== undefined) {
+        let event = await Promise.resolve(generator.next(instance));
+
+        while (!event.done) {
+          event = await Promise.resolve(generator.next(instance));
+        }
       }
+
+      return Promise.resolve(events);
+    };
+
+    for (const handler of handlerList) {
+      await _apply(handler);
     }
 
-    return Promise.resolve(events);
+    return events;
   };
 
   function rehydrate(events: IDomainEvent[], snapshot?: IAggregateState<T>) {
@@ -77,10 +89,10 @@ export function createAggregateRoot<T>(
   }
 
   return {
+    applyCommand,
     applyEvent,
     initialState,
     rehydrate,
-    applyCommand: handle,
     commands: Object.keys(commands).map(cmd => cmd.toLowerCase()),
     name: aggregateName.toLowerCase()
   };