+++ src/__tests__/AggregateRoot.spec.ts
@@ -1,6 +1,9 @@
 import test from 'ava';
+import Joi from 'joi';
+
 import { createAggregateRoot } from '../AggregateRoot';
-import { createCommand } from '../Command';
+import { createCommand, createCommandValidator } from '../Command';
+import { CommandValidationError, DomainError } from '../errors';
 import { createEvent } from '../Event';
 import { IAggregateDefinition } from '../interfaces';
 
@@ -16,18 +19,47 @@ const definition: IAggregateDefinition<ICounter> = {
       entity = yield entity.publish('incremented');
       yield entity.publish('incrementedBy', { step: entity.state.value });
     },
-    async *delayedIncrement(entity, _) {
+    arrayIncrements: [
+      (entity, _) => {
+        entity.publish('incremented');
+      },
+      (entity, _) => {
+        entity.publish('incremented');
+      },
+      (entity, _) => {
+        entity.publish('incrementedBy', { step: entity.state.value + 2 });
+      }
+    ],
+    async delayedIncrement(entity, { data }) {
+      await timeout(25);
+      if (data && data.step && data.step < 0) {
+        throw new DomainError('Increment step must be postiive');
+      }
+      entity.publish('incremented');
+    },
+    async *delayedYieldingIncrement(entity, _) {
       await timeout(25);
       yield entity.publish('incremented');
     },
     increment(entity, _) {
       entity.publish('incremented');
     },
-    *incrementByDynamic(entity, command) {
-      for (const step of command.data.steps) {
-        yield entity.publish('incrementedBy', { step });
+    incrementByPositive(entity, { data: { step } }) {
+      if (step <= 0) {
+        throw new DomainError('Must increment by positive number');
       }
-    }
+      entity.publish('incrementedBy', { step });
+    },
+    incrementByDynamic: [
+      createCommandValidator({
+        steps: Joi.array().items(Joi.number().integer())
+      }),
+      function* incrementByDynamic(entity, command) {
+        for (const step of command.data.steps) {
+          yield entity.publish('incrementedBy', { step });
+        }
+      }
+    ]
   },
   initialState: {
     value: 0
@@ -58,6 +90,14 @@ test('simple command handler', async t => {
   t.is(events.length, 1);
 });
 
+test('handle domain error', async t => {
+  const command = createCommand('incrementByPositive', 0, { step: -2 });
+  const shouldThrow = () =>
+    counterAggregate.applyCommand(counterAggregate.initialState, command);
+
+  await t.throwsAsync(shouldThrow, { name: 'DomainError' });
+});
+
 test('multiple yielding command handler', async t => {
   const incrementedByDynamic = createCommand('incrementByDynamic', 0, {
     steps: [1, 2, 3, 4]
@@ -75,6 +115,16 @@ test('multiple yielding command handler', async t => {
     createEvent('incrementedBy', { step: 3 }),
     createEvent('incrementedBy', { step: 4 })
   ]);
+});
+
+test('command validation', async t => {
+  const badCommand = createCommand('incrementByDynamic', 0, { steps: 'foo' });
+  const shouldThrow = () =>
+    counterAggregate.applyCommand(counterAggregate.initialState, badCommand);
+
+  await t.throwsAsync(shouldThrow, {
+    instanceOf: CommandValidationError
+  });
 });
 
 test('stateful multiple yielding', async t => {
@@ -97,11 +147,41 @@ test('stateful multiple yielding', async t => {
   ]);
 });
 
-test('asynchronous yielding', async t => {
+test('asynchronous commands', async t => {
   const initial = counterAggregate.initialState;
   const delayedIncrement = createCommand('delayedIncrement', 0);
   const events = await counterAggregate.applyCommand(initial, delayedIncrement);
 
   t.is(events.length, 1);
   t.deepEqual(events, [createEvent('incremented')]);
+});
+
+test('asynchronous error handling', async t => {
+  const initial = counterAggregate.initialState;
+  const delayedIncrement = createCommand('delayedIncrement', 0, { step: -1 });
+
+  const shouldThrow = counterAggregate.applyCommand(initial, delayedIncrement);
+
+  await t.throwsAsync(shouldThrow, { instanceOf: DomainError });
+});
+
+test('asynchronous yielding command handler', async t => {
+  const initial = counterAggregate.initialState;
+  const delayedIncrement = createCommand('delayedYieldingIncrement', 0);
+  const events = await counterAggregate.applyCommand(initial, delayedIncrement);
+
+  t.is(events.length, 1);
+  t.deepEqual(events, [createEvent('incremented')]);
+});
+
+test('multiple command handlers', async t => {
+  const initial = counterAggregate.initialState;
+  const arrayIncrement = createCommand('arrayIncrements', 0);
+  const events = await counterAggregate.applyCommand(initial, arrayIncrement);
+  t.is(events.length, 3);
+  t.deepEqual(events, [
+    createEvent('incremented'),
+    createEvent('incremented'),
+    createEvent('incrementedBy', { step: 4 })
+  ]);
 });